custom_nodes/requirements.txt
requests
pytrends
google-generativeai
pinecone-client
sentence-transformers
GitPython